moved controller from client to controller
