+++ dan-stineback/app/js/animals/controller/index.js
@@ -1,0 +1,5 @@
+'use strict';
+module.exports = function(app) {
+  require('./cat-controller')(app);
+  require('./dog-controller')(app);
+};
