+++ dan-stineback/app/js/animals/directives/index.js
