+++ dan-stineback/app/js/animals/directives/item-directive.js
