+++ dan-stineback/app/js/animals/directives/list-directive.js
