--- dan-stineback/app/js/directives/index.js
@@ -1,1 +1,0 @@
-'use strict';
