--- dan-stineback/app/js/directives/item-directive.js
@@ -1,1 +1,0 @@
-'use strict';
