--- dan-stineback/app/js/directives/list-directive.js
@@ -1,1 +1,0 @@
-'use strict';
